Add validations to login
