+++ client/src/pages/auth/components/register-form/register-form-general/RegisterFormGeneral.tsx
@@ -1,14 +1,16 @@
 import { SubmitHandler, useForm } from 'react-hook-form';
 import FormInput from '../../../../../components/common/form-input/FormInput';
 import {
+  EMAIL_VALIDATIONS,
+  PASSWORD_VALIDATIONS,
+} from '../../../validations-common';
+import {
   IGeneralStepper,
   IGeneralStepperForm,
   StepperButtonEnum,
 } from '../types';
 import {
-  EMAIL_VALIDATIONS,
   FIRST_NAME_VALIDATIONS,
-  PASSWORD_VALIDATIONS,
   REPEAT_PASSWORD_VALIDATIONS,
   SUR_NAME_VALIDATIONS,
 } from './validations';
